dispose of derived signals sooner
Or, to be more precise: if you have a signal that is scheduled for
disposal from another signal's disposer, it will now be disposed
synchronously, instead of a waiting for the next tick.

This was a goofy little behavior that we had before, that would
cause signal chains to clean up one tick at a time. Doing it all at
once is not only faster, but more predictable and expected.